requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
base58==2.1.1
beautifulsoup4==4.12.2
bitarray==2.8.1
cachetools==4.2.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
cytoolz==0.12.2
decorator==4.4.2
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.55.0
django-defender==0.9.7
django-redis==5.3.0
django-unicorn==0.54.0
eth-abi==4.2.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
frozendict==2.3.8
frozenlist==1.4.0
hexbytes==0.3.1
idna==3.4
ipfshttpclient==0.8.0a2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
multiaddr==0.0.9
multidict==6.0.4
netaddr==0.8.0
oauthlib==3.2.2
orjson==3.9.5
parsimonious==0.9.0
Pillow==10.0.0
protobuf==4.24.2
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.8.0
python-dateutil==2.7.5
python-dotenv==1.0.0
python3-openid==3.2.0
pyunormalize==15.0.0
pywin32==306
redis==5.0.0
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rlp==3.0.0
rpds-py==0.9.2
shortuuid==1.0.11
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
toolz==0.12.0
typing_extensions==4.3.0
tzdata==2023.3
urllib3==1.26.16
varint==1.0.2
web3==6.9.0
websockets==11.0.3
whitenoise==6.4.0
yarl==1.9.2
